wit/bindgen: consolidate testdata test code, writeFile for dumping bad Go code to disk
